Rename fruit search stuff
