+++ app/src/components/legend.jsx
