+++ app/stylesheets/main.css
@@ -21,7 +21,7 @@ body {
   background-color: rgb(70, 200, 180) !important;
 }
 
-.ProductTable {
+.ProductList {
   list-style: none;
   margin: 1em;
   padding: 1em;
@@ -53,7 +53,7 @@ body {
   font-family: monaco, Consolas, 'Lucida Console', monospace;
 }
 
-.ProductTable li {
+.ProductList li {
   margin: 0 .5em 1em;
   padding: 1em;
   width: 200px;
